% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{boston_water}
\alias{boston_water}
\title{Boston Water}
\format{
An sf dataframe with 10 observations
}
\usage{
data("boston_water")
}
\description{
This data contains the largest named water within Suffolk County MA, with geographies simplified.
}
\examples{
data('boston_water')
}
\concept{data}
