% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_water.R
\name{clip_water}
\alias{clip_water}
\title{Remove Water}
\usage{
clip_water(shp, water, filter_islands = FALSE, ...)
}
\arguments{
\item{shp}{input shp with \code{sf} geometry.}

\item{water}{water shapes to remove with \code{sf} geometry}

\item{filter_islands}{logical. Should additional filtering be done to remove small areas?}

\item{...}{additional arguments to pass to \code{rmapshaper::ms_filter_islands()}.
Only used if \code{filter_islands = TRUE}.}
}
\value{
tibble with \code{sf} geometries
}
\description{
Remove Water
}
\examples{
\donttest{
# time to run varies greatly, depending on machine
data(suffolk)
data(boston_water)
clip_water(suffolk, boston_water[10, ])
}
}
