% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dots.R
\name{create_dots}
\alias{create_dots}
\title{Creation of "dots" dataset}
\usage{
create_dots(num = 10)
}
\arguments{
\item{num}{The number of dots in the dataset}
}
\value{
A dataset
}
\description{
Creates a dataset with a specified number of dots to use with other dots functions
}
\examples{
dots_data = create_dots(num=10)

}
