% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{dots-package}
\alias{dots-package}
\title{Simplifying function calls}
\description{
Simplify your function definitions to reduce the complexity for the majority
of users while also providing venues for experienced users to do the fine
level control.
}
\author{
Leonardo Collado-Torres <lcollado@jhu.edu>
}
\keyword{package}

