% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_sf_regular}
\alias{engine_sf_regular}
\title{Generate Regular Points with \code{sf}}
\usage{
engine_sf_regular(shp, col)
}
\arguments{
\item{shp}{input shp with \code{sf} geometry.}

\item{col}{character column name to produce points with}
}
\value{
tibble with \code{sf} geometries
}
\description{
Uses \code{sf::st_sample()} to produce points and spatial joins with input \code{shp}.
Each engine function takes the same arguments and produces comparable outputs.
Final number of points may be approximate for this method.
}
\examples{
set.seed(1)
data('suffolk')
engine_sf_regular(suffolk[16:20,], 'pop_nhpi')
}
