% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_sp_hexagonal}
\alias{engine_sp_hexagonal}
\title{Generate Hexagonal Points with \code{sp}}
\usage{
engine_sp_hexagonal(shp, col)
}
\arguments{
\item{shp}{input shp with \code{sf} geometry.}

\item{col}{character column name to produce points with}
}
\value{
tibble with \code{sf} geometries
}
\description{
Uses \code{sp::spsample()} with method "hexagonal" to produce points, converts back to \code{sf},
and spatial joins with input \code{shp}.
Each engine function takes the same arguments and produces comparable outputs.
Final number of points may be approximate for this method.
}
\examples{
set.seed(1)
data('suffolk')
engine_sp_hexagonal(suffolk[16:20, ], 'pop_nhpi')
}
