% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine_terra}
\alias{engine_terra}
\title{Generate Points with \code{terra}}
\usage{
engine_terra(shp, col)
}
\arguments{
\item{shp}{input shp with \code{sf} geometry.}

\item{col}{character column name to produce points with}
}
\value{
tibble with \code{sf} geometries
}
\description{
Uses \code{terra::dots()} to produce points and transforms back to \code{sf}.
Each engine function takes the same arguments and produces comparable outputs.
}
\examples{
set.seed(1)
data('suffolk')
engine_terra(suffolk, 'pop_nhpi')
}
