% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pts.R
\name{filter_pts}
\alias{filter_pts}
\title{Filter Points to a Region}
\usage{
filter_pts(pts, shp, cond = TRUE)
}
\arguments{
\item{pts}{points with \code{sf} geometry to filter}

\item{shp}{shp to filter to}

\item{cond}{geometry subset to reduce shp to}
}
\value{
tibble with \code{sf} geometries
}
\description{
Filter Points to a Region
}
\examples{
data(suffolk)
pts <- dots_points(suffolk, pop, divisor = 1000)
filter_pts(pts, suffolk, pop < 1000)
}
