% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dots.r
\name{get_dots}
\alias{get_dots}
\title{An alternative way to access dots arguments without explicitly passing it through calling stack that allows updating default dots arguments that are explicitly set throughout calling stack.}
\usage{
get_dots(
  function_or_arg_list = NULL,
  select_args = NULL,
  search_while_calls_have_args = "...",
  search_while_calls_belong_to_env = NULL,
  search_while_calls_regexp = NULL,
  search_up_nframes = 1L,
  search_up_to_call = NULL
)
}
\arguments{
\item{function_or_arg_list}{The end function that meant to accept dots arguments (default arguments accessed with \code{formals(function_or_arg_list)}) or just explicit list of default dots arguments that will be searched up in calling stack and updated if set explicitly in higher calls.}

\item{select_args}{Which arguments to select from \code{function_or_arg_list}. Ether character or numeric vector.}

\item{search_while_calls_have_args}{Arguments that should be present in each upper call to continue looking up the call stack for updates in dots arguments.}

\item{search_while_calls_belong_to_env}{Environment/package name (character string) to which each function in upper calls to should belong to continue looking up the call stack for updates in dots arguments.}

\item{search_while_calls_regexp}{Regular expression that each function name in upper calls to should matched to continue looking up the call stack for updates in dots arguments.}

\item{search_up_nframes}{Number of frames (aka environments) in calling stack to look up for updates in dots arguments.}

\item{search_up_to_call}{The name of the call before which to continue looking up the call stack for updates in dots arguments.}
}
\value{
List of updated dots arguments
}
\description{
An alternative way to access dots arguments without explicitly passing it through calling stack that allows updating default dots arguments that are explicitly set throughout calling stack.
}
