% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_dots.R
\name{inner_function}
\alias{inner_function}
\title{Example: an inner function}
\usage{
inner_function(x, y = "old_y_default", z = "z_default")
}
\arguments{
\item{x}{A parameter}

\item{y}{A parameter with a default}

\item{z}{Another parameter with a default}
}
\description{
Along with \code{\link[=outer_function]{outer_function()}}, this function provides an example of the use
of \code{copy_dots()} in a package. See \code{\link[=outer_function]{outer_function()}}.
}
\examples{

## See outer_function()

}
\seealso{
\code{\link[=copy_dots]{copy_dots()}}

Other example: 
\code{\link{outer_function}()}
}
\concept{example}
