% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_dots.R
\name{move_dots}
\alias{move_dots}
\title{Movement of "dots"}
\usage{
move_dots(data, size = 100, movement = 5)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{size}{The size of the space within which to place the dots. Must be the same as the one used to place the dots.}

\item{movement}{The maximum distance that dots can move during one generation}
}
\value{
An updated dataset
}
\description{
Randomly moves dots within the delimited space
}
\examples{
dots_data = move_dots(data=dots_data, size=100, movement=5)

}
