% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{outer_function}
\alias{outer_function}
\title{Example: an outer function}
\usage{
outer_function(y = "new_y_default", ..., x, z = "z_default")
}
\arguments{
\item{y}{A parameter with a default}

\item{...}{Parameters passed to \code{inner_function()}}

\item{x}{A parameter}

\item{z}{Another parameter with a default}
}
\description{
Along with \code{\link[=inner_function]{inner_function()}}, this function provides an example of the use
of \code{copy_dots()} in a package. Ellipsis parameters (\code{...}) from this function
are passed to \code{\link[=inner_function]{inner_function()}}, but \code{copy_dots()} is unsed in the function
definition so that those parameters appear as parameters in the definition
and documentation of \code{\link[=outer_function]{outer_function()}}.
}
\examples{

inner_function(x = "inner_x_value")
outer_function(x = "outer_x_value")

}
\seealso{
\code{\link[=copy_dots]{copy_dots()}}

Other example: 
\code{\link{inner_function}()}
}
\concept{example}
