% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_dots.R
\name{place_dots}
\alias{place_dots}
\title{Placement of "dots"}
\usage{
place_dots(data, size = 100)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{size}{The size of the space within which to place the dots}
}
\value{
An updated dataset
}
\description{
Places the dots randomly within a delimited space
}
\examples{
dots_data = place_dots(data=dots_data, size=100)

}
