% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dots.R
\name{plot_dots}
\alias{plot_dots}
\title{Plots the "dots"}
\usage{
plot_dots(data, size = 100)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{size}{The size of the space within which to place the dots. Must be the same as the one used to place the dots.}
}
\value{
A plot
}
\description{
Plots the current position and status of the dots
}
\examples{
plot_dots(data=dots_data, size=100)

}
