% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_dots.R
\name{seed_dots}
\alias{seed_dots}
\title{Seeding of infected "dots"}
\usage{
seed_dots(data, numinf = 1)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{numinf}{The number of dots to infect}
}
\value{
An updated dataset
}
\description{
Randomly determines infected dots
}
\examples{
dots_data = seed_dots(data=dots_data, numinf=1)

}
