% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Create the transmission kernel parameters}
\usage{
set_params(size = 250, numinf = 2, movement = 10, inf_dist = 25,
  recover = 14, beta = 0.5, power = 0.9, offset = 10)
}
\arguments{
\item{size}{The size of the space within which the dots can move.}

\item{numinf}{The number of dots to infect}

\item{movement}{The maximum distance that dots can move during one generation}

\item{inf_dist}{The maximum distance that an infected dot can transmit the disease}

\item{recover}{The average number of generations before recovering after infection}

\item{beta}{Transmission parameter}

\item{power}{Kernel parameter}

\item{offset}{Kernel parameter}
}
\value{
A vector containing the parameter values
}
\description{
This is needed to execute the update_dots and simulate_dots functions
}
\examples{
params = set_params()

}
