% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dots.R
\name{simulate_dots}
\alias{simulate_dots}
\title{Executes a complete epidemic simulation with the "dots"}
\usage{
simulate_dots(data, p, cont_plot = T)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{p}{Transmission kernel parameters}

\item{cont_plot}{Logical. If TRUE, the plot will be updated for every generation. If FALSE, only a plot of the beginning and of the end of the epidemic will be generated.}
}
\value{
An updated dataset and plots of the dots.
}
\description{
This will randomly place and seed the dots before simulating an epidemic.
}
\details{
The epidemic will end either when all the dots are infected, or when there are no more infected.
}
\examples{
dots_data = simulate_dots(data=dots_data, p=params, cont_plot=FALSE)

}
