% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suffolk}
\alias{suffolk}
\title{Suffolk County, MA Voting Districts}
\format{
An sf dataframe with 295 observations
}
\usage{
data("suffolk")
}
\description{
This data contains the voting districts for Suffolk County MA, with geographies simplified.
}
\examples{
data('suffolk')
}
\concept{data}
