% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dots.R
\name{update_dots}
\alias{update_dots}
\title{Updating the status of the "dots"}
\usage{
update_dots(data, gen = 0, p)
}
\arguments{
\item{data}{The dataset containing the dots information}

\item{gen}{Current generation of the epidemic}

\item{p}{Transmission kernel parameters. These must have been created using the set_params function.}
}
\value{
An updated dataset
}
\description{
Updates the status of the dots (Susceptible, Infected or Recovered)
}
\examples{
dots_data = update_dots(data=dots_data, gen=0, p=params)

}
