% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dots.r
\docType{package}
\name{dots-package}
\alias{dots}
\alias{dots-package}
\title{dots: Gets ... in R function calls}
\description{
What the package does (one paragraph).
}
\details{
An alternative way to interact with ~...~ dots arguments (aka ellipses).

Provides access to ~...~ dots arguments without explicitly passing it through calling stack and allows updating default values that are explicitly set throughout calling stack (lower calls take prevalence).
}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
