% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.r
\docType{package}
\name{dots-package}
\alias{dots}
\alias{dots-package}
\title{dots: Provides access to arguments of nested functions. Sort of an alterative mechanism to passing `...` arguments but with more features.}
\description{
Provides access to higher level call's arguments (including `...` dots arguments) without explicitly passing it through calling stack and allows updating default values that are explicitly set throughout calling stack (i.e., lower calls take prevalence).
}
\details{
Provides access to arguments of nested functions. Sort of an alterative mechanism to passing \code{...} arguments but with more features.
Provides access to higher level call's arguments (including \code{...} dots arguments) without explicitly passing it through calling stack and allows updating default values that are explicitly set throughout calling stack (i.e., lower calls take prevalence).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stasvlasov.github.io/dots/}
  \item Report bugs at \url{https://github.com/stasvlasov/dots/issues}
}

}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
