% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dots.r
\docType{package}
\name{dots-package}
\alias{dots}
\alias{dots-package}
\title{dots: Gets ... in R function calls}
\description{
What the package does (one paragraph).
}
\details{
An alternative way to access dots arguments without explicitly passing it through calling stack that allows updating default dots arguments that are explicitly set throughout calling stack.
}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
